Move block sync module folder from startup/sync-block-task to block-sync-renderer
