+++ packages/neuron-wallet/package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "start": "electron .",
     "start:dev": "yarn run build && electron .",
-    "build": "ttsc && ncp ./src/startup/sync-block-task/index.html ./dist/startup/sync-block-task/index.html",
+    "build": "ttsc && ncp ./src/block-sync-renderer/index.html ./dist/block-sync-renderer/index.html",
     "clean": "rimraf dist/*",
     "test": "jest --runInBand",
     "test:e2e": "jest --config jest.e2e.config.js --runInBand",
